Add select model components for AI model management
- Introduced new components for selecting AI models, including SelectModel, LastUsedLlmModel, SelectSttModel, and SelectTtsModel.
- Created constants for provider options and defined types for model selection.
- Implemented fetching logic for various AI models from the API, enhancing user experience in selecting models.
- Integrated model selection into the chat features, allowing users to easily manage and switch between different AI models.
