+++ src/pages/user/features/ChatLlmList/component/selectModel/SelectSttModel.tsx
@@ -3,83 +3,83 @@ import axiosCustom from "../../../../../../config/axiosCustom";
 import Select from "react-select";
 import { LucideExpand, LucidePlus } from "lucide-react";
 
-export type TtsModelProvider = "openai" | "groq" | "localai" | "";
+export type SttModelProvider = "groq" | "openai" | "localai" | "";
 
-const TTS_PROVIDER_OPTIONS: { label: string; value: TtsModelProvider }[] = [
+const STT_PROVIDER_OPTIONS: { label: string; value: SttModelProvider }[] = [
     { label: 'Default (auto)', value: '' },
-    { label: 'OpenAI', value: 'openai' },
     { label: 'GROQ', value: 'groq' },
+    { label: 'OpenAI', value: 'openai' },
     { label: 'LocalAI', value: 'localai' },
 ];
 
-const TTS_DEFAULT_MODELS: Record<string, string> = {
-    openai: 'tts-1',
-    groq: 'canopylabs/orpheus-v1-english',
-    localai: '',
+const STT_DEFAULT_MODELS: Record<string, string> = {
+    groq: 'whisper-large-v3',
+    openai: 'whisper-1',
+    localai: 'whisper-tiny',
 };
 
-interface SelectTtsModelProps {
-    ttsModelProvider: string;
-    setTtsModelProvider: React.Dispatch<React.SetStateAction<string>>;
-    ttsModelName: string;
-    setTtsModelName: React.Dispatch<React.SetStateAction<string>>;
+interface SelectSttModelProps {
+    sttModelProvider: string;
+    setSttModelProvider: React.Dispatch<React.SetStateAction<string>>;
+    sttModelName: string;
+    setSttModelName: React.Dispatch<React.SetStateAction<string>>;
 }
 
-const SelectTtsModel: React.FC<SelectTtsModelProps> = ({
-    ttsModelProvider,
-    setTtsModelProvider,
-    ttsModelName,
-    setTtsModelName,
+const SelectSttModel: React.FC<SelectSttModelProps> = ({
+    sttModelProvider,
+    setSttModelProvider,
+    sttModelName,
+    setSttModelName,
 }) => {
     const [localaiModels, setLocalaiModels] = useState<Array<{ modelName: string; modelLabel: string; modelType?: string }>>([]);
     const [isLoadingLocalai, setIsLoadingLocalai] = useState(false);
-    const [expanded, setExpanded] = useState(true);
+    const [expanded, setExpanded] = useState(false);
 
     useEffect(() => {
-        if (ttsModelProvider !== 'localai') return;
+        if (sttModelProvider !== 'localai') return;
 
         const fetchLocalaiModels = async () => {
             setIsLoadingLocalai(true);
             try {
                 const res = await axiosCustom.get("/api/dynamic-data/model-localai/modelLocalaiGet");
                 const docs = res.data.docs || [];
-                const ttsModels = docs.filter((m: { modelType?: string }) =>
-                    m.modelType === 'tts' || m.modelType === ''
+                const sttModels = docs.filter((m: { modelType?: string }) =>
+                    m.modelType === 'stt' || m.modelType === ''
                 );
-                setLocalaiModels(ttsModels);
+                setLocalaiModels(sttModels);
 
-                if (ttsModels.length > 0) {
-                    setTtsModelName((prev) => {
-                        const valid = ttsModels.some((m: { modelName: string }) => m.modelName === prev);
-                        return valid ? prev : ttsModels[0].modelName;
+                if (sttModels.length > 0) {
+                    setSttModelName((prev) => {
+                        const valid = sttModels.some((m: { modelName: string }) => m.modelName === prev);
+                        return valid ? prev : sttModels[0].modelName;
                     });
                 }
             } catch (err) {
-                console.error('Failed to fetch LocalAI TTS models:', err);
+                console.error('Failed to fetch LocalAI STT models:', err);
             } finally {
                 setIsLoadingLocalai(false);
             }
         };
         fetchLocalaiModels();
-    }, [ttsModelProvider, setTtsModelName]);
+    }, [sttModelProvider, setSttModelName]);
 
     useEffect(() => {
-        if (ttsModelProvider === 'openai' || ttsModelProvider === 'groq') {
-            setTtsModelName(TTS_DEFAULT_MODELS[ttsModelProvider] || '');
-        } else if (ttsModelProvider === '') {
-            setTtsModelName('');
+        if (sttModelProvider === 'groq' || sttModelProvider === 'openai') {
+            setSttModelName(STT_DEFAULT_MODELS[sttModelProvider] || '');
+        } else if (sttModelProvider === '') {
+            setSttModelName('');
         }
-    }, [ttsModelProvider, setTtsModelName]);
+    }, [sttModelProvider, setSttModelName]);
 
     return (
         <Fragment>
             <div className="mb-2 lg:mb-3">
                 <div className="flex justify-between items-center">
                     <label className="block text-sm font-medium text-gray-700 mb-1 lg:mb-2">
-                        TTS (Text-to-Speech) - 
-                        {(ttsModelProvider && ttsModelName) ? (
+                        STT (Speech-to-Text) -
+                        {(sttModelProvider && sttModelName) ? (
                             <span className="text-gray-500">
-                                {ttsModelProvider} - {ttsModelName}
+                                {sttModelProvider} - {sttModelName}
                             </span>
                         ) : (
                             <span className="text-gray-500">
@@ -99,21 +99,22 @@ const SelectTtsModel: React.FC<SelectTtsModelProps> = ({
                 </div>
                 {expanded && (
                     <div className="flex flex-col sm:flex-row gap-2 mt-2">
-                        {TTS_PROVIDER_OPTIONS.map((opt) => (
+                        {STT_PROVIDER_OPTIONS.map((opt) => (
                             <button
                                 key={opt.value}
+                                type="button"
                                 onClick={() => {
-                                    setTtsModelProvider(opt.value);
+                                    setSttModelProvider(opt.value);
                                 }}
                                 className={
                                     `flex-1 px-3 py-2 text-sm rounded-sm border transition-colors
-                                    ${ttsModelProvider === opt.value
+                                    ${sttModelProvider === opt.value
                                         ? 'bg-blue-600 text-white border-blue-600'
                                         : 'bg-white text-gray-700 border-gray-300 hover:bg-blue-50'
                                     }
                                     font-semibold`
                                 }
-                                aria-pressed={ttsModelProvider === opt.value}
+                                aria-pressed={sttModelProvider === opt.value}
                             >
                                 {opt.label}
                             </button>
@@ -122,32 +123,32 @@ const SelectTtsModel: React.FC<SelectTtsModelProps> = ({
                 )}
             </div>
 
-            {expanded && ttsModelProvider === 'localai' && (
+            {expanded && sttModelProvider === 'localai' && (
                 <div className="mb-1 lg:mb-2">
-                    <h3 className="text-sm font-medium text-gray-700 mb-1 lg:mb-2">TTS Model</h3>
+                    <h3 className="text-sm font-medium text-gray-700 mb-1 lg:mb-2">STT Model</h3>
                     {isLoadingLocalai ? (
                         <div className="text-sm text-gray-500">Loading models...</div>
                     ) : localaiModels.length === 0 ? (
-                        <div className="text-sm text-gray-500">No TTS models found. Add models in LocalAI Settings.</div>
+                        <div className="text-sm text-gray-500">No STT models found. Add models in LocalAI Settings.</div>
                     ) : (
                         <Select<{ value: string; label: string }>
-                            value={ttsModelName ? { value: ttsModelName, label: localaiModels.find(m => m.modelName === ttsModelName)?.modelLabel || ttsModelName } : undefined}
-                            onChange={(opt) => opt && setTtsModelName(opt.value)}
+                            value={sttModelName ? { value: sttModelName, label: localaiModels.find(m => m.modelName === sttModelName)?.modelLabel || sttModelName } : undefined}
+                            onChange={(opt) => opt && setSttModelName(opt.value)}
                             options={localaiModels.map(m => ({ value: m.modelName, label: m.modelLabel || m.modelName }))}
-                            placeholder="Select TTS model..."
+                            placeholder="Select STT model..."
                             isSearchable={true}
                         />
                     )}
                 </div>
             )}
 
-            {expanded && (ttsModelProvider === 'openai' || ttsModelProvider === 'groq') && (
+            {expanded && (sttModelProvider === 'groq' || sttModelProvider === 'openai') && (
                 <p className="text-xs text-gray-500 mb-2">
-                    Using {TTS_DEFAULT_MODELS[ttsModelProvider]}.
+                    Using {STT_DEFAULT_MODELS[sttModelProvider]}.
                 </p>
             )}
         </Fragment>
     );
 };
 
-export default SelectTtsModel;
+export default SelectSttModel;
