+++ src/pages/user/features/ChatLlmList/component/selectModel/index.ts
@@ -1,0 +1,6 @@
+export { default as SelectModel } from './SelectModel';
+export { default as SelectSttModel } from './SelectSttModel';
+export { default as SelectTtsModel } from './SelectTtsModel';
+export type { ChatAiModelProvider, SelectModelProps } from './selectModel.types';
+export type { SttModelProvider } from './SelectSttModel';
+export type { TtsModelProvider } from './SelectTtsModel';
