+++ src/pages/user/features/ChatLlmList/component/selectModel/providers/SelectAiModelOpenrouter.tsx
@@ -1,0 +1,95 @@
+import { useState, useEffect } from "react";
+import Select from "react-select";
+import axiosCustom from "../../../../../../../config/axiosCustom";
+import { tsSchemaAiModelListOpenrouter } from "../../../../../../../types/pages/settings/dataModelOpenrouter";
+
+const SelectAiModelOpenrouter = ({
+    aiModelName,
+    setAiModelName,
+    selectRandomModel,
+}: {
+    aiModelName: string;
+    setAiModelName: React.Dispatch<React.SetStateAction<string>>;
+    selectRandomModel: number;
+}) => {
+    const [modelArr, setModelArr] = useState([] as tsSchemaAiModelListOpenrouter[]);
+    const [isLoadingModel, setIsLoadingModel] = useState(true);
+
+    useEffect(() => {
+        const fetchModelData = async () => {
+            try {
+                setIsLoadingModel(true);
+                const response = await axiosCustom.get('/api/dynamic-data/model-openrouter/modelOpenrouterGet');
+
+                if (response.data.docs && response.data.docs.length > 0) {
+                    let tempModelArr = response.data.docs as {
+                        id: string;
+                        name: string;
+                        description: string;
+                    }[];
+
+                    tempModelArr = tempModelArr.map((model) => ({
+                        id: model.id,
+                        name: `${model.name} (${model.id})`,
+                        description: model.description,
+                    })).sort((a, b) => a.name.localeCompare(b.name));
+
+                    if (aiModelName === '') {
+                        if (tempModelArr.length > 0) {
+                            setAiModelName(tempModelArr[0].id);
+                        }
+                    }
+
+                    setModelArr(tempModelArr);
+                }
+            } catch (error) {
+                console.error('Error fetching model data:', error);
+            } finally {
+                setIsLoadingModel(false);
+            }
+        };
+        fetchModelData();
+    }, []);
+
+    useEffect(() => {
+        if (selectRandomModel >= 1) {
+            if (modelArr.length > 0) {
+                const randomModel = modelArr[Math.floor(Math.random() * modelArr.length)];
+                setAiModelName(randomModel.id);
+            }
+        }
+    }, [selectRandomModel]);
+
+    useEffect(() => {
+        if (aiModelName === '') {
+            if (modelArr.length > 0) {
+                setAiModelName(modelArr[0].id);
+            }
+        }
+    }, [aiModelName]);
+
+    return (
+        <div className="mb-2">
+            <h3 className="text-sm font-medium text-gray-700 mb-2">OpenRouter Model</h3>
+            <Select<{ value: string; label: string }>
+                value={aiModelName ? { value: aiModelName, label: modelArr.find(model => model.id === aiModelName)?.name || "" } : undefined}
+                onChange={(selectedOption: { value: string; label: string } | null) => {
+                    if (selectedOption) {
+                        setAiModelName(selectedOption.value);
+                    }
+                }}
+                options={
+                    modelArr.map((model) => ({
+                        value: model.id,
+                        label: `${model.name} (${model.id})`
+                    }))
+                }
+                placeholder="Select a model..."
+                isLoading={isLoadingModel}
+                isSearchable={true}
+            />
+        </div>
+    );
+};
+
+export default SelectAiModelOpenrouter;
