+++ src/pages/user/features/ChatLlmList/sectionRightChat/ThreadAdd/ComponentThreadAdd.tsx
@@ -1,475 +1,13 @@
 import { Link, useNavigate } from "react-router-dom";
 import axiosCustom from "../../../../../../config/axiosCustom";
 import toast from "react-hot-toast";
-import { MessageCircle, Settings, ExternalLink, LucideInfo } from "lucide-react";
-import SelectSttModel from "../ThreadSetting/SelectSttModel";
-import SelectTtsModel from "../ThreadSetting/SelectTtsModel";
+import { MessageCircle, LucideInfo } from "lucide-react";
+import { SelectModel } from "../../component/selectModel";
 import { useState, useEffect } from "react";
-import Select from "react-select";
-import { tsSchemaAiModelListGroq } from "../../../../../../types/pages/settings/dataModelGroq";
-import { tsSchemaAiModelListOpenrouter } from "../../../../../../types/pages/settings/dataModelOpenrouter";
 import { jotaiChatThreadRefreshRandomNum } from "../../jotai/jotaiChatLlmThreadSetting";
 import { useAtomValue, useSetAtom } from "jotai";
 import Tooltip from '@rc-component/tooltip';
 import stateJotaiAuthAtom from "../../../../../../jotai/stateJotaiAuth";
-
-const SelectAiModelOpenrouter = ({
-    aiModelName,
-    setAiModelName,
-
-    selectRandomModel,
-}: {
-    aiModelName: string;
-    setAiModelName: React.Dispatch<React.SetStateAction<string>>;
-
-    selectRandomModel: number;
-}) => {
-    const [modelArr, setModelArr] = useState([] as tsSchemaAiModelListOpenrouter[]);
-    const [isLoadingModel, setIsLoadingModel] = useState(true);
-
-    useEffect(() => {
-        const fetchModelData = async () => {
-            try {
-                setIsLoadingModel(true);
-                const response = await axiosCustom.get('/api/dynamic-data/model-openrouter/modelOpenrouterGet');
-
-                if (response.data.docs && response.data.docs.length > 0) {
-
-                    let tempModelArr = response.data.docs as {
-                        id: string;
-                        name: string;
-                        description: string;
-                    }[];
-
-                    tempModelArr = tempModelArr.map((model) => ({
-                        id: model.id,
-                        name: `${model.name} (${model.id})`,
-                        description: model.description,
-                    })).sort((a, b) => a.name.localeCompare(b.name));
-
-                    // if aiModelName is empty, select a random model
-                    if (aiModelName === '') {
-                        if (tempModelArr.length > 0) {
-                            setAiModelName(tempModelArr[0].id);
-                        }
-                    }
-
-                    setModelArr(tempModelArr);
-                }
-            } catch (error) {
-                console.error('Error fetching model data:', error);
-                // Keep default model if API fails
-            } finally {
-                setIsLoadingModel(false);
-            }
-        };
-        fetchModelData();
-    }, []);
-
-    useEffect(() => {
-        if (selectRandomModel >= 1) {
-            if (modelArr.length > 0) {
-                const randomModel = modelArr[Math.floor(Math.random() * modelArr.length)];
-                setAiModelName(randomModel.id);
-            }
-        }
-    }, [selectRandomModel]);
-
-    useEffect(() => {
-        if (aiModelName === '') {
-            if (modelArr.length > 0) {
-                setAiModelName(modelArr[0].id);
-            }
-        }
-    }, [aiModelName]);
-
-    return (
-        <div className="mb-2">
-            <h3 className="text-sm font-medium text-gray-700 mb-2">OpenRouter Model</h3>
-            <Select<{ value: string; label: string }>
-                value={aiModelName ? { value: aiModelName, label: modelArr.find(model => model.id === aiModelName)?.name || "" } : undefined}
-                onChange={(selectedOption: { value: string; label: string } | null) => {
-                    if (selectedOption) {
-                        setAiModelName(selectedOption.value);
-                    }
-                }}
-                options={
-                    modelArr.map((model: any) => ({
-                        value: model.id,
-                        label: `${model.name} (${model.id})`
-                    }))
-                }
-
-                placeholder="Select a model..."
-                isLoading={isLoadingModel}
-                isSearchable={true}
-            />
-        </div>
-    )
-}
-
-const SelectAiModelGroq = ({
-    aiModelName,
-    setAiModelName,
-
-    selectRandomModel,
-}: {
-    aiModelName: string;
-    setAiModelName: React.Dispatch<React.SetStateAction<string>>;
-
-    selectRandomModel: number;
-}) => {
-    const [modelArr, setModelArr] = useState([] as tsSchemaAiModelListGroq[]);
-    const [isLoadingModel, setIsLoadingModel] = useState(true);
-
-    useEffect(() => {
-        const fetchModelData = async () => {
-            setIsLoadingModel(true);
-            const response = await axiosCustom.get('/api/dynamic-data/model-groq/modelGroqGet');
-            setModelArr(response.data.docs);
-
-            // if aiModelName is empty, select a random model
-            if (aiModelName === '') {
-                if (modelArr.length > 0) {
-                    setAiModelName(modelArr[0].id);
-                }
-            }
-
-            setIsLoadingModel(false);
-        }
-        fetchModelData();
-    }, []);
-
-    useEffect(() => {
-        if (selectRandomModel >= 1) {
-            if (modelArr.length > 0) {
-                const randomModel = modelArr[Math.floor(Math.random() * modelArr.length)];
-                setAiModelName(randomModel.id);
-            }
-        }
-    }, [selectRandomModel]);
-
-    useEffect(() => {
-        if (aiModelName === '') {
-            if (modelArr.length > 0) {
-                setAiModelName(modelArr[0].id);
-            }
-        }
-    }, [aiModelName]);
-
-    return (
-        <div className="mb-2">
-            <h3 className="text-sm font-medium text-gray-700 mb-2">GROQ Model</h3>
-            <Select<{ value: string; label: string }>
-                value={aiModelName ? { value: aiModelName, label: modelArr.find(model => model.id === aiModelName)?.id || "" } : undefined}
-                onChange={(selectedOption: { value: string; label: string } | null) => {
-                    if (selectedOption) {
-                        setAiModelName(selectedOption.value);
-                    }
-                }}
-                options={
-                    modelArr.map((model: any) => ({
-                        value: model.id,
-                        label: `${model.owned_by} - ${model.id}`
-                    }))
-                }
-
-                placeholder="Select a model..."
-                isLoading={isLoadingModel}
-                isSearchable={true}
-            />
-        </div>
-    )
-}
-
-const SelectAiModelOllama = ({
-    aiModelName,
-    setAiModelName,
-}: {
-    aiModelName: string;
-    setAiModelName: React.Dispatch<React.SetStateAction<string>>;
-}) => {
-    const [ollamaModels, setOllamaModels] = useState<
-        Array<{ _id: string; modelName: string; modelLabel: string }>
-    >([]);
-    const [loading, setLoading] = useState<boolean>(false);
-    const [error, setError] = useState<string>("");
-
-    useEffect(() => {
-        const fetchOllamaModels = async () => {
-            setLoading(true);
-            setError("");
-            try {
-                const res = await axiosCustom.get("/api/dynamic-data/model-ollama/modelOllamaGet");
-                setOllamaModels(res.data.docs || []);
-            } catch (err) {
-                setError("Failed to fetch models");
-            } finally {
-                setLoading(false);
-            }
-        };
-        fetchOllamaModels();
-    }, []);
-
-    useEffect(() => {
-        // Set default to first model if not set
-        if (ollamaModels.length > 0 && !aiModelName) {
-            setAiModelName(ollamaModels[0].modelName);
-        }
-    }, [ollamaModels, aiModelName, setAiModelName]);
-
-    return (
-        <div className="mb-2">
-            <h3 className="text-sm font-medium text-gray-700 mb-2">Ollama Model</h3>
-            {loading ? (
-                <div className="text-gray-500 text-sm">Loading models...</div>
-            ) : error ? (
-                <div className="text-red-600 text-sm">{error}</div>
-            ) : ollamaModels.length === 0 ? (
-                <div className="text-sm text-gray-500">No Ollama models found. Go to Ollama settings to add models.</div>
-            ) : (
-                <select
-                    className="block w-full px-3 py-2 border border-gray-300 rounded-md text-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                    value={aiModelName}
-                    onChange={e => setAiModelName(e.target.value)}
-                >
-                    {ollamaModels.map(model => (
-                        <option key={model._id} value={model.modelName}>
-                            {model.modelLabel || model.modelName}
-                        </option>
-                    ))}
-                </select>
-            )}
-        </div>
-    );
-};
-
-const SelectAiModelLocalai = ({
-    aiModelName,
-    setAiModelName,
-}: {
-    aiModelName: string;
-    setAiModelName: React.Dispatch<React.SetStateAction<string>>;
-}) => {
-    const [localaiModels, setLocalaiModels] = useState<
-        Array<{ _id: string; modelName: string; modelLabel: string }>
-    >([]);
-    const [loading, setLoading] = useState<boolean>(false);
-    const [error, setError] = useState<string>("");
-
-    useEffect(() => {
-        const fetchLocalaiModels = async () => {
-            setLoading(true);
-            setError("");
-            try {
-                const res = await axiosCustom.get("/api/dynamic-data/model-localai/modelLocalaiGet");
-                setLocalaiModels(res.data.docs || []);
-            } catch (err) {
-                setError("Failed to fetch models");
-            } finally {
-                setLoading(false);
-            }
-        };
-        fetchLocalaiModels();
-    }, []);
-
-    useEffect(() => {
-        // Set default to first model if not set
-        if (localaiModels.length > 0 && !aiModelName) {
-            setAiModelName(localaiModels[0].modelName);
-        }
-    }, [localaiModels, aiModelName, setAiModelName]);
-
-    return (
-        <div className="mb-2">
-            <h3 className="text-sm font-medium text-gray-700 mb-2">LocalAI Model</h3>
-            {loading ? (
-                <div className="text-gray-500 text-sm">Loading models...</div>
-            ) : error ? (
-                <div className="text-red-600 text-sm">{error}</div>
-            ) : localaiModels.length === 0 ? (
-                <div className="text-sm text-gray-500">No LocalAI models found. Go to LocalAI settings to add models.</div>
-            ) : (
-                <select
-                    className="block w-full px-3 py-2 border border-gray-300 rounded-md text-sm focus:outline-none focus:ring-blue-500 focus:border-blue-500"
-                    value={aiModelName}
-                    onChange={e => setAiModelName(e.target.value)}
-                >
-                    {localaiModels.map(model => (
-                        <option key={model._id} value={model.modelName}>
-                            {model.modelLabel || model.modelName}
-                        </option>
-                    ))}
-                </select>
-            )}
-        </div>
-    );
-};
-
-const SelectAiModelOpenaiCompatible = ({
-    aiModelName,
-    setAiModelName,
-    aiModelOpenAiCompatibleConfigId,
-    setAiModelOpenAiCompatibleConfigId,
-}: {
-    aiModelName: string;
-    setAiModelName: React.Dispatch<React.SetStateAction<string>>;
-    aiModelOpenAiCompatibleConfigId: string | null;
-    setAiModelOpenAiCompatibleConfigId: React.Dispatch<React.SetStateAction<string | null>>;
-}) => {
-    interface IOpenaiCompatibleModel {
-        _id: string;
-        providerName?: string;
-        baseUrl: string;
-        modelName?: string;
-    }
-
-    const [configs, setConfigs] = useState<IOpenaiCompatibleModel[]>([]);
-    const [isLoadingModel, setIsLoadingModel] = useState(true);
-
-    useEffect(() => {
-        const fetchConfigs = async () => {
-            setIsLoadingModel(true);
-            try {
-                const response = await axiosCustom.post<{ docs: IOpenaiCompatibleModel[] }>(
-                    `/api/user/openai-compatible-model/crud/openaiCompatibleModelGet`,
-                    {},
-                    {
-                        headers: {
-                            'Content-Type': 'application/json'
-                        },
-                        withCredentials: true,
-                    }
-                );
-                setConfigs(response.data.docs || []);
-
-                // Auto-select first config if none selected
-                if (!aiModelOpenAiCompatibleConfigId && response.data.docs && response.data.docs.length > 0) {
-                    const firstConfig = response.data.docs[0];
-                    setAiModelOpenAiCompatibleConfigId(firstConfig._id);
-                    setAiModelName(firstConfig.modelName || '');
-                }
-            } catch (error) {
-                console.error('Error fetching OpenAI compatible models:', error);
-            } finally {
-                setIsLoadingModel(false);
-            }
-        };
-        fetchConfigs();
-    }, []);
-
-    return (
-        <div className="mb-2">
-            <h3 className="text-sm font-medium text-gray-700 mb-1 lg:mb-2">OpenAI Compatible Model</h3>
-            {isLoadingModel ? (
-                <div className="text-sm text-gray-500">Loading configurations...</div>
-            ) : configs.length === 0 ? (
-                <div className="text-sm text-gray-500 mb-2">
-                    No configurations found.
-                    <Link to="/user/setting/openai-compatible-model" className="text-blue-600 hover:underline ml-1">
-                        Create one here
-                    </Link>
-                </div>
-            ) : (
-                <Select<{ value: string; label: string }>
-                    value={aiModelOpenAiCompatibleConfigId ? {
-                        value: aiModelOpenAiCompatibleConfigId,
-                        label: (() => {
-                            const config = configs.find(c => c._id === aiModelOpenAiCompatibleConfigId);
-                            if (!config) return '';
-                            const displayName = config.providerName || config.baseUrl;
-                            return aiModelName ? `${displayName} - ${aiModelName}` : displayName;
-                        })()
-                    } : undefined}
-                    onChange={(selectedOption: { value: string; label: string } | null) => {
-                        if (selectedOption) {
-                            const config = configs.find(c => c._id === selectedOption.value);
-                            setAiModelOpenAiCompatibleConfigId(selectedOption.value);
-                            setAiModelName(config?.modelName || '');
-                        }
-                    }}
-                    options={configs.map((config) => {
-                        const displayName = config.providerName || config.baseUrl;
-                        const label = config.modelName ? `${displayName} - ${config.modelName}` : displayName;
-                        return { value: config._id, label };
-                    })}
-                    placeholder="Select a configuration..."
-                    isSearchable={true}
-                />
-            )}
-        </div>
-    );
-};
-
-const LastUsedLlmModel = ({
-    aiModelName,
-    setAiModelName,
-    setAiModelProvider,
-}: {
-    aiModelName: string;
-    setAiModelName: React.Dispatch<React.SetStateAction<string>>;
-    setAiModelProvider: React.Dispatch<React.SetStateAction<"openrouter" | "groq" | "ollama" | "localai" | "openai-compatible">>;
-}) => {
-    const [topLlmModels, setTopLlmModels] = useState<Array<{ aiModelProvider: string, aiModelName: string }>>([]);
-
-    const [showMoreModels, setShowMoreModels] = useState(false);
-
-    useEffect(() => {
-        const fetchTopLlmModels = async () => {
-            try {
-                const response = await axiosCustom.get('/api/chat-llm/threads-crud/topLlmConversationModel');
-                setTopLlmModels(response.data.modelArr || []);
-            } catch (error) {
-                console.error('Error fetching top LLM models:', error);
-            }
-        };
-
-        fetchTopLlmModels();
-    }, []);
-
-    const handleModelSelect = (model: { aiModelProvider: string, aiModelName: string }) => {
-        if (model.aiModelProvider === 'openrouter' || model.aiModelProvider === 'groq' || model.aiModelProvider === 'ollama' || model.aiModelProvider === 'localai' || model.aiModelProvider === 'openai-compatible') {
-            setAiModelProvider(model.aiModelProvider as "openrouter" | "groq" | "ollama" | "localai" | "openai-compatible");
-        }
-        setAiModelName(model.aiModelName);
-    };
-
-    // Helper function to format model name for display
-    const formatModelNameForDisplay = (modelName: string): string => {
-        // For all providers, just format slashes
-        // modelName now contains only the model name (no configId prefix)
-        return modelName.replace('/', ' / ').replace('/', ' / ');
-    };
-
-    return (
-        <div className="mb-2">
-            {topLlmModels.slice(0, showMoreModels ? topLlmModels.length : 4).map((model, index) => (
-                <button
-                    key={index}
-                    onClick={() => handleModelSelect(model)}
-                    className={`text-left px-2 py-1 text-sm rounded-sm border hover:bg-gray-50 ${aiModelName === model.aiModelName ? 'bg-blue-50 border-blue-200' : 'border-gray-200'
-                        }`}
-                >
-                    <div>
-                        <span className="inline-block text-gray-700 text-sm">
-                            {formatModelNameForDisplay(model?.aiModelName || '')}
-                        </span>
-                    </div>
-                </button>
-            ))}
-
-            {topLlmModels.length > 4 && (
-                <button
-                    onClick={() => setShowMoreModels(!showMoreModels)}
-                    className="text-sm text-gray-500 hover:text-gray-700 inline-block border border-gray-200 px-2 py-1 rounded"
-                >
-                    <span className="mr-1">🔍</span>
-                    {showMoreModels ? 'Show Less' : 'Show More'}
-                </button>
-            )}
-        </div>
-    );
-}
 
 const ComponentThreadAdd = () => {
     const navigate = useNavigate();
@@ -632,172 +170,24 @@ const ComponentThreadAdd = () => {
                     Configure the assistant once. You can change models anytime in thread settings.
                 </p>
 
-                <div className="mb-5 rounded-2xl border border-zinc-200/80 bg-white/90 p-4 shadow-lg shadow-zinc-900/[0.04] ring-1 ring-black/[0.02] backdrop-blur-sm sm:p-5">
-                    <div className="mb-3 flex items-center justify-between">
-                        <h3 className="text-xs font-semibold uppercase tracking-wider text-zinc-500">
-                            Model
-                        </h3>
-                        <button
-                            type="button"
-                            className="text-zinc-400 hover:text-zinc-700"
-                            title="Settings"
-                        >
-                            <Settings className="h-4 w-4" />
-                        </button>
-                    </div>
-
-                    <div className="mb-2">
-                        <h3 className="mb-1.5 text-xs font-medium text-zinc-700">Provider</h3>
-                        <div className="mb-2 flex flex-col gap-1.5 sm:flex-row sm:flex-wrap">
-                            {[
-                                { label: 'OpenRouter', value: 'openrouter' },
-                                { label: 'GROQ', value: 'groq' },
-                                { label: 'Ollama', value: 'ollama' },
-                                { label: 'LocalAI', value: 'localai' },
-                                { label: 'OpenAI Compatible', value: 'openai-compatible' }
-                            ].map((provider) => (
-                                <button
-                                    key={provider.value}
-                                    onClick={() => {
-                                        setAiModelProvider(provider.value as "openrouter" | "groq" | "ollama" | "localai" | "openai-compatible");
-                                        setAiModelName('');
-                                        setAiModelOpenAiCompatibleConfigId(null);
-                                        setSelectRandomModel(Math.floor(Math.random() * 1000000));
-                                    }}
-                                    className={`flex-1 rounded-xl border px-2 py-2 text-xs font-medium transition-all sm:min-w-[5.5rem] ${
-                                        aiModelProvider === provider.value
-                                            ? 'border-teal-500/30 bg-gradient-to-br from-teal-600 to-emerald-600 text-white shadow-md shadow-teal-900/15'
-                                            : 'border-zinc-200/90 bg-zinc-50/80 text-zinc-800 hover:border-zinc-300 hover:bg-white'
-                                    }`}
-                                    aria-pressed={aiModelProvider === provider.value}
-                                >
-                                    {provider.label}
-                                </button>
-                            ))}
-                        </div>
-                    </div>
-
-                    {/* field -> select model -> openrouter */}
-                    {aiModelProvider === 'openrouter' && (
-                        <SelectAiModelOpenrouter
-                            aiModelName={aiModelName}
-                            setAiModelName={setAiModelName}
-                            selectRandomModel={selectRandomModel}
-                            key={'select-model-openrouter'}
-                        />
-                    )}
-
-                    {/* field -> select model -> groq */}
-                    {aiModelProvider === 'groq' && (
-                        <SelectAiModelGroq
-                            aiModelName={aiModelName}
-                            setAiModelName={setAiModelName}
-                            selectRandomModel={selectRandomModel}
-                            key={'select-model-groq'}
-                        />
-                    )}
-
-                    {/* field -> select model -> ollama */}
-                    {aiModelProvider === 'ollama' && (
-                        <div>
-                            <SelectAiModelOllama
-                                aiModelName={aiModelName}
-                                setAiModelName={setAiModelName}
-                                key={'select-model-ollama'}
-                            />
-
-                            <div className="text-sm text-gray-500 mt-2">
-                                Manage your Ollama models in the{' '}
-                                <Link
-                                    to="/user/setting/ollama-models"
-                                    className="text-blue-500 hover:text-blue-700"
-                                >Ollama Settings</Link>.
-                            </div>
-                        </div>
-                    )}
-
-                    {/* field -> select model -> localai */}
-                    {aiModelProvider === 'localai' && (
-                        <div>
-                            <SelectAiModelLocalai
-                                aiModelName={aiModelName}
-                                setAiModelName={setAiModelName}
-                                key={'select-model-localai'}
-                            />
-
-                            <div className="text-sm text-gray-500 mt-2">
-                                Manage your LocalAI models in the{' '}
-                                <Link
-                                    to="/user/setting/localai-models"
-                                    className="text-blue-500 hover:text-blue-700"
-                                >LocalAI Settings</Link>.
-                            </div>
-                        </div>
-                    )}
-
-                    {/* field -> select model -> openai-compatible */}
-                    {aiModelProvider === 'openai-compatible' && (
-                        <div>
-                            <SelectAiModelOpenaiCompatible
-                                aiModelName={aiModelName}
-                                setAiModelName={setAiModelName}
-                                aiModelOpenAiCompatibleConfigId={aiModelOpenAiCompatibleConfigId}
-                                setAiModelOpenAiCompatibleConfigId={setAiModelOpenAiCompatibleConfigId}
-                                key={'select-model-openai-compatible'}
-                            />
-
-                            <div className="text-sm text-gray-500 mt-2">
-                                Manage your OpenAI compatible model configurations in the{' '}
-                                <Link
-                                    to="/user/setting/openai-compatible-model"
-                                    className="text-blue-500 hover:text-blue-700"
-                                >OpenAI Compatible Model Settings</Link>.
-                            </div>
-                        </div>
-                    )}
-
-                    {/* field -> select model -> llm */}
-                    <LastUsedLlmModel
-                        aiModelName={aiModelName}
-                        setAiModelProvider={setAiModelProvider}
-                        setAiModelName={setAiModelName}
-                        key={'select-model-last-used-llm'}
-                    />
-
-                    {/* field -> sttModelProvider */}
-                    <SelectSttModel
-                        sttModelProvider={sttModelProvider}
-                        setSttModelProvider={setSttModelProvider}
-                        sttModelName={sttModelName}
-                        setSttModelName={setSttModelName}
-                    />
-
-                    {/* field -> ttsModelProvider */}
-                    <SelectTtsModel
-                        ttsModelProvider={ttsModelProvider}
-                        setTtsModelProvider={setTtsModelProvider}
-                        ttsModelName={ttsModelName}
-                        setTtsModelName={setTtsModelName}
-                    />
-
-                    {/* field -> buttons */}
-                    <div className="mt-2">
-                        <Link to="/user/setting" className="text-sm text-gray-500 hover:text-gray-700 inline-block mr-5">
-                            <ExternalLink className="w-4 h-4 mr-1 inline-block" />
-                            Model page
-                        </Link>
-
-                        <button
-                            onClick={() => {
-                                setSelectRandomModel(selectRandomModel + 1);
-                            }}
-                            className="mt-2 text-sm text-blue-500 hover:text-blue-700 inline-block"
-                        >
-                            <span className="mr-1">🎲</span>
-                            Random LLM
-                        </button>
-                    </div>
-                </div>
+                <SelectModel
+                    aiModelProvider={aiModelProvider}
+                    setAiModelProvider={setAiModelProvider}
+                    aiModelName={aiModelName}
+                    setAiModelName={setAiModelName}
+                    aiModelOpenAiCompatibleConfigId={aiModelOpenAiCompatibleConfigId}
+                    setAiModelOpenAiCompatibleConfigId={setAiModelOpenAiCompatibleConfigId}
+                    sttModelProvider={sttModelProvider}
+                    setSttModelProvider={setSttModelProvider}
+                    sttModelName={sttModelName}
+                    setSttModelName={setSttModelName}
+                    ttsModelProvider={ttsModelProvider}
+                    setTtsModelProvider={setTtsModelProvider}
+                    ttsModelName={ttsModelName}
+                    setTtsModelName={setTtsModelName}
+                    selectRandomModel={selectRandomModel}
+                    setSelectRandomModel={setSelectRandomModel}
+                />
 
                 <div className="mb-5 rounded-2xl border border-zinc-200/80 bg-white/90 p-4 shadow-lg shadow-zinc-900/[0.04] ring-1 ring-black/[0.02] backdrop-blur-sm sm:p-5">
                     <div className="flex flex-col lg:flex-row lg:space-x-8 space-y-4 lg:space-y-0">
